Added Catalan support (TODO: improve interpunct handling)
Added the Catalan alphabet:
"ca": "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyzÀÈÉÍÏÒÓÚÜÇàèéíïòóúüç"

There is, though, one special case that is a bit tricky: the interpunct, "·". In Catalan, "l·l" (sometimes misspelled as "l.l" or "l•l") belongs to the alphabet, but the rest of the letters with "·" don't. For instance, "r·r" does not belong to the alphabet. Provisionally, perhaps we could consider "·" as another character of the alphabet?